Refactor graph logic into modules
